support combining publishToMavenCentral with properties for automatic release and validation
